# docs/swagger.yaml
basePath: /
definitions:
  customer.AppointmentResponse:
    properties:
      amount:
        type: integer
      appointment_date:
        type: string
      created_at:
        type: string
      end_time:
        type: string
      id:
        type: integer
      merchant_name:
        type: string
      order_no:
        type: string
      service_name:
        type: string
      staff_name:
        type: string
      start_time:
        type: string
      status:
        type: string
    type: object
...
  customer.CreateAppointmentRequest:
    properties:
      coupon_id:
        description: 可选
        type: integer
      date:
        type: string
      merchant_id:
        type: integer
      remark:
        type: string
      service_id:
        type: integer
      staff_id:
        type: integer
      time_slot_id:
        type: integer
    required:
    - date
    - merchant_id
    - service_id
    - staff_id
    - time_slot_id
    type: object
...
  customer.SlotResponse:
    properties:
      end_time:
        type: string
      id:
        type: integer
      start_time:
        type: string
    type: object
...
  models.Service:
    properties:
      categoryID:
        type: integer
      coverImage:
        type: string
      createdAt:
        type: string
      description:
        type: string
      duration:
        description: 分钟
        type: integer
      id:
        type: integer
      isActive:
        type: boolean
      merchantID:
        type: integer
      name:
        type: string
      price:
        type: integer
      sort:
        type: integer
      updatedAt:
        type: string
    type: object
...
  models.Staff:
    properties:
      avatar:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      merchantID:
        type: integer
      name:
        type: string
      position:
        type: string
      specialties:
        type: string
      updatedAt:
        type: string
    type: object
...
  utils.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 预约系统API文档
  title: Appointment System API
  version: "1.0"
paths:
  /api/customer/appointments:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的预约记录，支持按状态筛选
      parameters:
      - description: 预约状态筛选
        enum:
        - pending
        - confirmed
        - completed
        - cancelled
        in: query
        name: status
        type: string
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回预约列表
          schema:
            items:
              $ref: '#/definitions/customer.AppointmentResponse'
            type: array
        "500":
          description: 获取预约列表失败
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取用户预约列表
      tags:
      - 预约管理
    post:
      consumes:
      - application/json
      description: 用户创建新的服务预约
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 预约创建请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/customer.CreateAppointmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回预约信息
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 参数错误或日期格式错误
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 创建预约失败
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建预约
      tags:
      - 预约管理
...
  /api/customer/merchants/{merchantId}/services:
    get:
      consumes:
      - application/json
      description: 根据商家ID获取服务列表，支持按分类ID筛选
      parameters:
      - description: 商家ID
        example: 123
        in: path
        name: merchantId
        required: true
        type: integer
      - description: 分类ID
        example: 456
        in: query
        name: categoryId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回服务列表
          schema:
            $ref: '#/definitions/models.Service'
        "400":
          description: 无效的商家ID
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 获取服务列表失败
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取商家服务列表
      tags:
      - 商家服务
...
  /api/customer/profile:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的个人信息（包括统计信息）
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回用户信息
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 获取用户信息失败
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取用户信息
      tags:
      - 用户管理
  /api/customer/services/{serviceId}/staff:
    get:
      consumes:
      - application/json
      description: 根据服务ID获取可提供该服务的员工列表
      parameters:
      - description: 服务ID
        example: 789
        in: path
        name: serviceId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回可用员工列表
          schema:
            items:
              $ref: '#/definitions/models.Staff'
            type: array
        "400":
          description: 无效的服务ID
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 获取可选员工失败
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取服务的可用员工
      tags:
      - 服务管理
  /api/customer/timeslots:
    get:
      consumes:
      - application/json
      description: 查询指定商家、技师在指定日期的可预约时间段
      parameters:
      - description: 商家ID
        in: query
        name: merchantId
        required: true
        type: integer
      - description: 技师ID
        in: query
        name: staffId
        required: true
        type: integer
      - description: '日期 (格式: YYYY-MM-DD)'
        example: "2023-06-15"
        in: query
        name: date
        required: true
        type: string
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回可预约时间段列表
          schema:
            items:
              $ref: '#/definitions/customer.SlotResponse'
            type: array
        "400":
          description: 参数错误（无效的商家ID、员工ID或日期格式）
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 获取时间段失败
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取可预约时间段
      tags:
      - 时间槽管理
  /api/customer/timeslots/dates:
    get:
      consumes:
      - application/json
      description: 查询指定商家、技师和服务的可预约日期范围（默认14天内）
      parameters:
      - description: 商家ID
        in: query
        name: merchantId
        required: true
        type: integer
      - description: 技师ID
        in: query
        name: staffId
        type: integer
      - description: 服务ID
        in: query
        name: serviceId
        type: integer
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回可预约日期列表（YYYY-MM-DD格式）
          schema:
            items:
              type: string
            type: array
        "400":
          description: 无效的商家ID
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 获取可预约日期失败
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取可预约日期列表
      tags:
      - 时间槽管理
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"